requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
pandas==2.0.3
gunicorn==21.2.0